refactor(ccw.core-hovers): Changed location of AbstractModelObject and the changed hover font name.
Small refactor of AbstractModelObject and the internal mapping of the hover font name/id.
