+++ ccw.core/src/java/ccw/core/StaticStrings.java
@@ -1,5 +1,7 @@
 package ccw.core;
 
+import org.eclipse.jface.resource.JFaceResources;
+
 import ccw.CCWPlugin;
 
 /**
@@ -24,8 +26,8 @@ public class StaticStrings {
 	// Files
 	public final static String CCW_HOVER_CSS = "platform:/plugin/" + CCWPlugin.PLUGIN_ID + "/assets/hover-control-style.css";
 
-	// Symbolic Names
-	public final static String CCW_HOVER_FONT = org.eclipse.jdt.ui.PreferenceConstants.APPEARANCE_JAVADOC_FONT;
+	// Fonts
+	public final static String CCW_FONT_HOVER_DEFAULT = JFaceResources.DEFAULT_FONT;
 
 	// Colors
 	public final static String CCW_COLOR_REPL_EXPR_BACKGROUND = "ccw.editors.clojure.colors.repl.expressionBackground";
